% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_by_vectors.R
\name{align_by_vectors}
\alias{align_by_vectors}
\title{Aligns 3D data by two vectors}
\usage{
align_by_vectors(l1, l2, coords, e_coords)
}
\arguments{
\item{l1}{A 2x3 matrix-like object with coordinates defining a line, one per row.}

\item{l2}{A 2x3 matrix-like object with coordinates defining a line, one per row.}

\item{coords}{A Nx3 matrix-like object containing xyz point coordinates, one
per row. Normally these would be mesh vertices, obtained with t(mesh$vb)[,1:3]}

\item{e_coords}{A Nx3 matrix-like object containing xyz point coordinates,
one per row, for additional coordinates which should be kept separately (e.g.
landmarks).}
}
\value{
A list containing rotated input coordinates (coords and e_coords)
}
\description{
Rotates the input 3D data using two vectors, such that both
vectors are on planes parallel to the XY plane. The data will be rotated
such that the first vector (l1) will always have identical Y and Z values for
both endpoints, with Y=0, and the second vector (l2) will always have
endpoints with Z=0 and the first point at the origin (i.e. x=0,y=0,z=0)
}
\note{
CRITICAL NOTE: It is not guaranteed to align with consistent
left/right orientation.
}
\section{TODO}{
 Change the code to align direction as well (i.e. positive along x axis
from first pt to second.)
}

\examples{
data(demoFlake2)
len <- demoFlake2$lms[c(3,nrow(demoFlake2$lms)), ] # Pt. percussion to tip
pw <- demoFlake2$lms[c(1,nrow(demoFlake2$lms) - 1), ] # Platform width vec
\dontrun{
shade3d(demoFlake2$mesh, color="green")
points3d(len, color="red", size=10)
points3d(pw, color="blue", size=10)
}
# Orient mesh vertices and landmarks by length and PW vectors:
res <- align_by_vectors(len, pw,  t(demoFlake2$mesh$vb)[, 1:3], demoFlake2$lms)
len.rot <- res$e_coords[c(3,nrow(demoFlake2$lms)), ]
pw.rot <- res$e_coords[c(1,nrow(demoFlake2$lms) - 1), ]
\dontrun{
# Plot result in 2D to visualize rotation:
plot(res$coords, asp=1)
lines(len.rot, col="red", lwd=2)
lines(pw.rot, col="purple", lwd=2)
}
}
\author{
Cornel M. Pop
}
