% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_rotate.R
\name{alignAxis}
\alias{alignAxis}
\title{Align to axis (deprecated)}
\usage{
alignAxis(l, coords)
}
\arguments{
\item{l}{A 2X3 matrix-like object containing xyz point coordinates defining
the alignment vector, one coordinate per row.}

\item{coords}{A Nx3 matrix-like object containing xyz point coordinates, one
per row.}
}
\value{
A list containing rotated coordinates (coords) and the rotated
input vector (l).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is obsolete. Please use the more general
\code{\link[=rotate_v2v]{rotate_v2v()}} function instead.
}
\author{
Cornel M. Pop
}
\keyword{internal}
