% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_dist.R
\name{coords_onseg}
\alias{coords_onseg}
\title{Determine if coordinates are on a line segment}
\usage{
coords_onseg(coords, seg, tol = 0)
}
\arguments{
\item{coords}{A Nx3 matrix-like object with x, y, and z point coordinates,
one per row.}

\item{seg}{A 2x3 matrix-like object containing x, y, and z coordinates for
two points that define the line segment.}

\item{tol}{(optional) Tolerance value for distance from the line segment;
useful to deal with floating point precision errors. Default is 0.}
}
\value{
A Boolean vector indicating, for each input coordinate, whether they
lie on the given line segment
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Determines if the given input coordinates lie on a line segment
or not.
}
\examples{
coords = data.frame(x = c(14.830469, 12.812613),
                    y = c(-84.840050, -79.966871),
                    z = c(2.938235, 2.302472))
seg = data.frame(x = c(12.84073, 12.49499), y = c(-80.03477, -79.19981),
                 z = c(2.311330, 2.202399))
coords_onseg(coords, seg) # False, False
coords_onseg(coords, seg, tol = 0.001) # False, True

\dontrun{
library(rgl)
points3d(coords, color = "red")
lines3d(seg), color = "blue")
}

}
\seealso{
\code{\link[=proj_pt2l]{proj_pt2l()}} for projecting coordinates orthogonally onto a line.
}
\author{
Cornel M. Pop
}
