% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demoFlake2}
\alias{demoFlake2}
\title{Flake 2}
\format{
A list containing a mesh3d object (mesh) and a set of basic landmarks
(lms) defining the ventral surface of the flake
}
\source{
3D scan courtesy of Pop Geological Consulting Services
LTD \url{http://www.popgeology.com}
}
\usage{
data(demoFlake2)
}
\description{
A demo flake (PGCS_Vau.2) for use with the Lithics3D package.
}
\author{
Cornel M. Pop
}
\keyword{datasets}
