% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demoSphere}
\alias{demoSphere}
\title{Demo sphere}
\format{
A mesh3d object
}
\usage{
data(demoSphere)
}
\description{
A small closed sphere used to test functions within this package
}
\keyword{datasets}
