% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demoSurface}
\alias{demoSurface}
\title{Demo surface}
\format{
A mesh3d object
}
\usage{
data(demoSurface)
}
\description{
A small open surface used to test functions within this package
}
\keyword{datasets}
