% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_dist.R
\name{dist_pt2l}
\alias{dist_pt2l}
\title{Euclidean distance between a point and a line}
\usage{
dist_pt2l(coords, l)
}
\arguments{
\item{coords}{A Nx3 matrix-like object with point coordinates, one
per row.}

\item{l}{A 2x3 matrix-like object with coordinates defining a line, one
coordinate per row.}
}
\value{
A vector of distances.
}
\description{
Computes the minimum distance between points and a line

Computes minimum distances between a given set of 3D coordinates
and a line
}
\note{
This is a very simple distance function. The orthogonal projection
of the point onto the line is done by \link{proj_pt2l}
}
\section{TODO}{
 Tests
}

\author{
Cornel M. Pop
}
