% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_collinear.R
\name{.check.mp.d}
\alias{.check.mp.d}
\title{Test for distance between coordinate means and path midpoint}
\usage{
.check.mp.d(c.res, lms)
}
\arguments{
\item{c.res}{A NxK matrix-like object containing xyz point coordinates,
one per row.}

\item{lms}{A vector of length 3 containing the ids (row numbers) of the
coordinates to use for the test.}
}
\value{
Distance between the midpoint of the path and the mean of the given
coordinates, in mesh units.
}
\description{
Test for distance between coordinate means and path midpoint.
}
\section{TODO}{
 document this function better, and write tests. Change
parameter from c.res to coords, and lms to lm.id or something
}

\author{
Cornel M. Pop
}
\keyword{internal}
