% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeAngles_yezziwoodleyetal.R
\name{.cluster_patch}
\alias{.cluster_patch}
\title{Classify patch coordinates into one of two surfaces}
\usage{
.cluster_patch(coords, normals, rad, lambda)
}
\description{
Classify patch coordinates into one of two surfaces
}
\keyword{internal}
