% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_collinear.R
\name{.collin.check}
\alias{.collin.check}
\title{Silly collinearity check function}
\usage{
.collin.check(c, lms, thresh = NA)
}
\arguments{
\item{c}{An NxK matrix-like object containing xyz point coordinates,
one per row.}

\item{lms}{A vector of length 3 containing the ids (row numbers) of the
coordinates to use for the test.}

\item{thresh}{A numeric threshold, given in mesh units, specifying the
threshold distance below which points will be considered colinear. If NA
(default), only strict colinearity will be tested.}
}
\value{
Boolean indicating whether the points are collinear (True) or not
False).
}
\description{
A silly collinearity check auxiliary function.
}
\section{TODO}{
 Document better and write unit test?
}

\author{
Cornel M. Pop
}
