% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_collinear.R
\name{.p3.collin}
\alias{.p3.collin}
\title{Test for the colinearity of 3 3D points.}
\usage{
.p3.collin(coords)
}
\arguments{
\item{coords}{A NxK matrix-like object containing xy(z) point coordinates,
one per row. Columns k > 3 are not considered}
}
\value{
True if collinear, False if not.
}
\description{
A simple function to test for colinearity of points in 3D space
}
\section{TODO}{
 Write tests
}

\author{
Cornel M. Pop
}
\keyword{internal}
