% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeAngles_yezziwoodleyetal.R
\name{.PCA_smallest_eig}
\alias{.PCA_smallest_eig}
\title{Compute smallest eigenvalue and its corresponding eigenvector}
\usage{
.PCA_smallest_eig(xyz_data, C, c, center = FALSE)
}
\description{
Compute smallest eigenvalue and its corresponding eigenvector
}
\keyword{internal}
