% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeAngles_yw.R
\name{.withness}
\alias{.withness}
\title{Identify data split that minimizes variance}
\usage{
.withness(x)
}
\arguments{
\item{x}{A numeric vector containing the data to be analyzed.}
}
\value{
A list containing the following components:
\itemize{
\item \code{w}: The minimum within-group sum of squares (variance) for
the optimal split.
\item \code{m}: The value at the optimal split point in the sorted data.
\item \code{mlow}: The lower boundary value of the primary cluster around
the split point.
\item \code{mhigh}: The upper boundary value of the primary cluster
around the split point.
}
}
\description{
This function evaluates a numeric vector to determine the split point that
minimizes the within-group variance.
}
\details{
This is a translation of c++ code from the Meshlab Virtual Goniometer plugin
developed by the \href{http://amaaze.umn.edu}{AMAAZE consortium}
}
\examples{
\dontrun{
data <- rnorm(100)
result <- Lithics3D:::.withness(data)
print(result)
}
}
\keyword{internal}
