% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_mark_pois.R
\name{drop_poi}
\alias{drop_poi}
\title{Drop the last point of interest (POI).}
\usage{
drop_poi(pois)
}
\arguments{
\item{pois}{A data frame output by the \code{\link[=mesh_mark_pois]{mesh_mark_pois()}}
function}
}
\value{
A reduced data frame
}
\description{
Removes the last POI from the given dataset, updates the current 3D scene
accordingly, and returns the modified dataset.
}
\examples{
\dontrun{
# Example usage:
# Create a dataset with marked POIs
mesh <- mesh_recenter(demoFlake2$mesh)$mesh
pois <- mesh_mark_pois(mesh)

# Drop the last POI and update the 3D scene
pois <- drop_poi(pois)
}

}
\seealso{
\code{\link[=mesh_mark_pois]{mesh_mark_pois()}} for interactive POI selection
}
\keyword{3D}
\keyword{interactive}
\keyword{landmark}
