% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_angles_valletta.R
\name{edge_angles_valletta}
\alias{edge_angles_valletta}
\title{Compute edge angles using Valletta et al.'s method}
\usage{
edge_angles_valletta(mesh, pois, radius = 3)
}
\arguments{
\item{mesh}{A triangular mesh (\code{mesh3d}) object on which edge angles will be
measured. Please ensure normals have been updated (see Examples).}

\item{pois}{Two points of interest (i.e., POIs) along the edge between which
the edge angle will be measured; represented as a 2X3 matrix-like object
containing x, y, and z coordinates. Please ensure the POIs are on the mesh
surface (see Details).}

\item{radius}{A positive numeric value indicating the search radius from the
line defined by the \code{pois} for including \code{mesh} vertices in the angle
measurements (see Details).}
}
\value{
A list containing:
\enumerate{
\item \code{angle}: a numeric value corresponding to the measured angle, in degrees.
\item \code{seg_1}: a 2x3 data.frame object with the x, y, and z coordinates (one
per row) of the end points of the first intersecting line segment
(perpendicular to the edge, and along the first intersecting surface).
\item \code{seg_2}: a 2x3 data.frame object with the x, y, and z coordinates (one
per row) of the end points of the second intersecting line segment
(perpendicular to the edge, and along the second intersecting surface).
\item \code{ip}: a 1x3 data.frame object with the x, y, and z coordinates of the
seg_1 and seg_2 intersection point. Corresponds to the input parameter \code{poi}.
\item \code{diag}: a list with other diagnostic data including:
a. \code{method}: a string with the name of the method used to measure the
angle (i.e., V2020).
b. \code{timestamp}: a string with the time when the angle was measured.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Implements the algorithm proposed by Valletta et al. (2020) for measuring
edge angles.
}
\note{
TODO: Add Details and specify why POI should be on the mesh surface, amongst
other things.

TODO:
\itemize{
\item test output to make sure it is correct (test against original method)
\item implement algorithm for removal of outliers and optimization (see page 49
of the publication)
}
}
\examples{
pois <- data.frame(x = c(-21.57720, -16.07506),
y = c(44.94063, 18.35520),
z = c(414.3339, 408.5220), Tag = c(1, 2))
mesh <- demoFlake2$mesh
mesh <- Rvcg::vcgUpdateNormals(mesh)
res <- edge_angles_valletta(mesh, pois, radius = 3)
}
\author{
Cornel M. Pop
}
