% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeAngles_visualize.R
\name{edge_angles_vis2d}
\alias{edge_angles_vis2d}
\title{2D visualizer for edgeAngles output}
\usage{
edge_angles_vis2d(ea_output, mesh, ncol = 2, nrow = 2)
}
\arguments{
\item{ea_output}{A list corresponding to the unmodified output of
\code{\link[=edgeAngles]{edgeAngles()}}.}

\item{mesh}{An \code{mesh3d} object corresponding to the mesh on which the angles
were measured.}

\item{ncol}{A number indicating the number of sub-plots to fit horizontally
in each output figure.}

\item{nrow}{A number indicating the number of sub-plots to fit vertically
in each output figure.}
}
\value{
A list of figures containing sub-plots for each angle measurement
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Generates figures from the output of the \code{\link[=edgeAngles]{edgeAngles()}} function.
}
\details{
The output figures include one or more sub-plots, each of which shows a
slice of the mesh and the lines (with endpoints) used to measure the edge
angles on that slice. Note that:
\itemize{
\item The mesh slices are perpendicular to the object edge, and on the
plane (\verb{P)} on which the angles were measured (see \code{\link[=edgeAngles]{edgeAngles()}} for more
details).
\item The black dots that define the mesh slice outlines
represent the endpoints of the triangle sides (i.e., mesh edges)
intersected by the measurement plane (\code{P}); in other words, they do not
indicate exact intersection points between the triangle sides and the
measurement plane - they are an approximation.
\item Data in each sub-plot has been rotated so that they align with the
x and y axes, with the origin (i.e., x = 0, y = 0) set to the object's edge
sampling point.
\item Axis values are given in mesh units.
\item Each sub-plot is labelled with the index of the respective angle
measurement in the \code{ea_output} list.
\item With each output figure all sub-plots are drawn to the same scale and with
a 1:1 ratio, but the scale may vary across figures.
\item The number of sub-plots per figure is determined from the \code{ncol} and \code{nrow}
parameters. If individual figures for each sub-plot are needed, simply set
both \code{ncol} and \code{nrow} to 1.
}
}
\examples{
\dontrun{
# Compute edge angles using included demo data:
data(demoFlake2)
e.curve = sPathConnect(demoFlake2$lms[1:4, ],
                       demoFlake2$mesh, path.choice = "ridges")
mv <- t(demoFlake2$mesh$vb)
path.res <- pathResample(mv[e.curve,1:3], 30, method = "npts")
res = edgeAngles(demoFlake2$mesh, path.res, m.d = 3)

# Generate figures:
angle_plots <- edge_angles_vis2d(res,
                                 mesh = demoFlake2$mesh,
                                 ncol = 3)
# See the first figure:
angle_plots[[1]]
}
}
\seealso{
\link{edge_angles_vis3d} - visualize the edge angle measurements in 3D.
}
\author{
Cornel M. Pop
}
