% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surf_edgeIntersects.R
\name{edgesOnPlane}
\alias{edgesOnPlane}
\title{Mesh edge / plane intersection}
\usage{
edgesOnPlane(p, mpts, medges)
}
\arguments{
\item{p}{A 3x3 matrix-like object with coordinates defining a plane, one per
row.}

\item{mpts}{An Nx4 matrix-like object corresponding to the transposed mesh$vb}

\item{medges}{An Nx4 dataframe corresponding to the output of
Rvcg::vcgGetEdge(mesh, unique=T).}
}
\value{
A vector containing row IDs in medges identifying edges intersecting
the given plane.
}
\description{
Identifies the mesh edges that are intersected by the given
plane. Note that no intersection points are computed by this function. For
that functionality use the Morpho::meshPlaneIntersect function.
}
\note{
The input of this function may be changed to simply be a mesh object,
once I write a new function to cache the calls (expected to be repeated).
}
\section{TODO}{
 Write memoised function for mpts and medges
}

\examples{
print("TODO")
}
