% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surf_Area.R
\name{getTArea}
\alias{getTArea}
\title{Compute triangle area from 3D coordinates}
\usage{
getTArea(coords)
}
\arguments{
\item{coords}{a vector of concatenated xyz coordinates. Note that only the
first 9 values will be used.}
}
\value{
numeric value of the computed area (0 for degenerate triangles) or
NA in case of bad input
}
\description{
Computes the area of a triangle from 3D coordinates
}
\note{
this function is currently too slow - 100k calls are processed in ~3.4
seconds.
}
\examples{
# input format: c(x,y,z,x1,y1,z1,x2,y2,z2)
triangle <- c(-2,0,5,-10,-2,-1,0,0,0)
Lithics3D:::getTArea(triangle)
}
\keyword{internal}
