% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_volume.R
\name{getTVol}
\alias{getTVol}
\title{Compute tetrahedron volume from triangle}
\usage{
getTVol(coords)
}
\arguments{
\item{coords}{a vector of concatenated xyz coordinates. Note that only the
first 9 values will be used.}
}
\value{
numeric value of the computed volume or NA in case of bad input
}
\description{
Computes the signed volume of a tetrahedron defined by the 3D
coordinates of a triangle (i.e. triangle vertices + origin as vertex)
}
\note{
this function is currently too slow - 100k calls are processed in ~5.2
seconds.
}
\examples{
# input format: c(x,y,z,x1,y1,z1,x2,y2,z2)
triangle <- c(-2,0,5,-10,-2,-1,1,0,0)
Lithics3D:::getTVol(triangle)
}
\keyword{internal}
