% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_l2lIntersect.R
\name{l2lIntersect}
\alias{l2lIntersect}
\title{Compute 2D line/line intersect}
\usage{
l2lIntersect(l1, l2)
}
\arguments{
\item{l1}{2D line coefficients}

\item{l2}{2D line coefficients}
}
\value{
A vector with 2D intersection coordinates or NA if lines are parallel
}
\description{
Compute line intersects based on line equation coefficients
This function is not meant to be called directly.
}
\note{
translated from rook at http://stackoverflow.com/questions/20677795/how-do-i-compute-the-intersection-point-of-two-lines-in-python
}
\examples{
l1 <- rbind(c(1,5), c(3,2))
l2 <- rbind(c(2,3), c(4,4))
# 2.250, 3.125
Lithics3D:::l2lIntersect(Lithics3D:::lineCoefs(l1[1,], l1[2,]),
                         Lithics3D:::lineCoefs(l2[1,], l2[2,]))
}
\author{
Cornel M. Pop
}
\keyword{internal}
