% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_l2sIntersect.R
\name{l2sIntersect}
\alias{l2sIntersect}
\title{Line/sphere intersection points}
\usage{
l2sIntersect(l, s)
}
\arguments{
\item{l}{A 2x3 matrix-like object with coordinates defining a line, one per
row.}

\item{s}{A vector of length 4 (center.x, center.y, center.z, radius) defining
a sphere.}
}
\value{
A list containing coordinates of intersection points, or NA if the
line does not intersect the sphere. If the line is tangent to the sphere,
both points will be identical.
}
\description{
Computes the intersection points of a line and a sphere
}
\note{
Formula taken from Fell, Harriet (2011) CS4300 class notes as accessed
at http://www.ccs.neu.edu/home/fell/CS4300/Lectures/Ray-TracingFormulas.pdf
}
\section{TODO}{
 Fix the description of the return
}

\examples{
library(rgl)
l = matrix(c(0.3,2.5,3,2,5,5), nrow=2, byrow=TRUE)
s = c(2,2,2,3)
res <- l2sIntersect(l, s)
spheres3d(s[1], s[2], s[3], radius=s[4], color="blue", alpha=0.4)
points3d(res, color="red")
lines3d(l, color="black", lwd=3)
}
\author{
Cornel M. Pop
}
