% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_l2lIntersect.R
\name{lineCoefs}
\alias{lineCoefs}
\title{Compute 2D line coefficients}
\usage{
lineCoefs(pt1, pt2)
}
\arguments{
\item{pt1}{a vector with 2D coordinates of a point on the line. Additional
dimensions are ignored}

\item{pt2}{a vector with 2D coordinates of a second point on the (direction).
Additional dimensions are ignored}
}
\value{
A vector with computed line equation coefficients
}
\description{
Computes line equation coefficients
This function is not meant to be called directly.
}
\note{
Translated from rook at http://stackoverflow.com/questions/20677795/how-do-i-compute-the-intersection-point-of-two-lines-in-python
}
\examples{
l <- rbind(c(1,5), c(3,2))
Lithics3D:::lineCoefs(l[1,], l[2,]) # 3, 2, 13
l <- rbind(c(1,5,2), c(3,2,0))
Lithics3D:::lineCoefs(l[1,], l[2,]) # 3, 2, 13
}
\author{
Cornel M. Pop
}
\keyword{internal}
