% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_l2lIntersect.R
\name{lineIntersect2D}
\alias{lineIntersect2D}
\title{Compute 2D line/line intersection}
\usage{
lineIntersect2D(l1, l2)
}
\arguments{
\item{l1}{a 2D coordinate matrix-like object defining two points (one per
row) along the first line. Only the first two columns (x,y) will be used.}

\item{l2}{a 2D coordinate matrix-like object defining two points (one per
row) along the second line. Only the first two columns (x,y) will be used.}
}
\value{
A vector with 2D intersection coordinates or NA if lines are parallel
}
\description{
Compute line intersects based on coordinates for points along
the lines
}
\note{
This is a wrapper function for the l2lIntersect, which is awkward to
use as is.
}
\examples{
l1 <- rbind(c(1,5), c(3,2))
l2 <- rbind(c(2,3), c(4,4))
lineIntersect2D(l1, l2) # 2.250, 3.125
}
\author{
Cornel M. Pop
}
