% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lithics3d-package.R
\name{Lithics3D}
\alias{Lithics3D}
\alias{Lithics3D-package}
\title{A toolbox for 3D analysis of archaeological lithics}
\description{
A collection of functions for working with and analyzing 3D scans of
archaeological lithics
}
\author{
Cornel M. Pop
}
