% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surf_Area.R
\name{mesh_area}
\alias{mesh_area}
\title{Computes mesh surface area}
\usage{
mesh_area(mesh, par_exec = FALSE, max_cores = NA)
}
\arguments{
\item{mesh}{a clean mesh3d object for which area will be computed}

\item{par_exec}{boolean indicating whether par_exec processing is to be used.
Set to FALSE by default}

\item{max_cores}{maximum number of cores to use if parallel processing is
requested. Requesting more cores than are physically available will simply
result in all cores being used.}
}
\value{
Numeric value of the computed area or NA in case of bad input.
}
\description{
Calculates the surface area of an arbitrary triangular
mesh by adding up the areas of individual constituent triangles. Please
ensure that the mesh is clean (e.g. no non-manifold vertices, etc)
}
\note{
For small meshes it may be faster to use a single thread due to the
overhead associated with starting the processing cluster.
}
\examples{
data(demoSurface)
# Non-parallel execution:
mesh_area(demoSurface)
# Parallel execution with a maximum of 2 cores.
mesh_area(demoSurface, par_exec = TRUE, max_cores = 2)
}
