% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_recenter.R
\name{mesh_recenter}
\alias{mesh_recenter}
\title{Recenter mesh coordinates}
\usage{
mesh_recenter(mesh, coords = NA)
}
\arguments{
\item{mesh}{An object of class \code{mesh3d} representing the 3D mesh.}

\item{coords}{(Optional) An Nx3 matrix-like object containing xyz
coordinates, one coordinate per row. Default is \code{NA}.}
}
\value{
A list containing the translated mesh (\code{mesh}) and coordinates
(\code{coords})
}
\description{
Transforms a mesh so that its center is at the origin (0, 0, 0);
it does so by subtracting the average x, y, and z coordinates from all
vertices. The same transformation can be applied to the optional set of
coordinates to preserve spatial relationships between the mesh and these
coordinates.
}
\section{TODO}{
 Add tests
}

\examples{
\dontrun{
library(rgl)
data(demoFlake2)
res <- mesh_recenter(demoFlake2$mesh, demoFlake2$lms)
view3d(theta=0, phi=0)
shade3d(res$mesh, color="green")
points3d(res$coords, color="red", size=5)
}

}
\author{
Cornel M. Pop
}
