% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_splitbypath.R
\name{mesh_segment_by_path}
\alias{mesh_segment_by_path}
\title{Section mesh using a path}
\usage{
mesh_segment_by_path(
  mesh,
  lms = NULL,
  path.choice = "ridges",
  mesh.path = NULL
)
}
\arguments{
\item{mesh}{A (clean) \code{mesh3d} object}

\item{lms}{(\emph{Deprecated}) An ordered set of coordinates. Since this parameter
is being deprecated, please use the output of \code{\link[=sPathConnect]{sPathConnect()}} to the
\code{mesh.path} argument)}

\item{path.choice}{(\emph{Deprecated}) Described in \code{\link[=sPathConnect]{sPathConnect()}}.}

\item{mesh.path}{A vector containing a list of connected vertices forming a
closed path on the surface of the input mesh. This parameter should
correspond to the output of \code{\link[=sPathConnect]{sPathConnect()}}}
}
\value{
A list of mesh objects containing a minimum of 3 faces each.
}
\description{
Cuts a mesh along a closed path defined on the basis of an
ordered set of coordinates (ideally landmarks), and splits the mesh into
multiple components. Note that you must provide a clean mesh, so please use
\code{\link[Rvcg:vcgClean]{Rvcg::vcgClean()}} on the input.
}
\note{
Minimum output mesh segment size is 3 faces, \strong{not including faces
along the mesh border}. Anything smaller gets discarded silently.

Note also that the deprecated parameters will stop working in version 1 of
this package at the latest.
}
\examples{
library(rgl)
mesh_path <- sPathConnect(demoFlake2$lms, demoFlake2$mesh,
                          path.choice = "ridges", closed = TRUE)
mesh_segs <- mesh_segment_by_path(demoFlake2$mesh, mesh.path = mesh_path)
\dontrun{
shade3d(mesh_segs[[1]], color = "green")
shade3d(mesh_segs[[2]], color = "blue")
close3d()
}
}
\author{
Cornel M. Pop
}
