% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_thicknessMap.R
\name{mesh_tmap}
\alias{mesh_tmap}
\title{Generate a thickness map of an oriented mesh}
\usage{
mesh_tmap(mesh.o, base.res, ld.cutoff)
}
\arguments{
\item{mesh.o}{An oriented mesh3d object}

\item{base.res}{Base resolution requested for the thickness map, in mesh
units. This resolution will be used only if the number of cells containing
less than 5 values is fewer than the limit specified by the ld.cutoff
parameter. Otherwise the resolution will be increased in .1 increments until
the condition is met.}

\item{ld.cutoff}{Threshold value (0-1) used to determine the maximum
percentage of grid cells that may have fewer than 5 mesh vertices assigned to
them. Above this value the resolution of the map will be lowered (i.e. the
resolution value will go up, resulting in larger cells).}
}
\description{
This generates a thickeness map of an un-split mesh (i.e. we
don't know the surfaces ) by assigning the mesh vertices to a grid of a given
resolution and measuring thickness in each grid cell.
}
\section{TODO}{
 Document properly what's going on here
}

