% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_thicknessMap.R
\name{mesh_tmap_plot}
\alias{mesh_tmap_plot}
\title{Plot thickness map}
\usage{
mesh_tmap_plot(tmap, font_size, y.label = "PC2 (mm)", x.label = "",
  annot = "A")
}
\arguments{
\item{tmap}{A thickness map as output by the \link{mesh_tmap} function}

\item{font_size}{A numeric value specifying the font size argument to be
passed on to \link{ggplot}}

\item{y.label}{A text to be used to label the y axis}

\item{x.label}{A text to be used to label the x axis}

\item{annot}{An annotation to be added to the resulting graph}
}
\value{
A ggplot object, ready to be plotted.
}
\description{
Plots a given thickness map with some sensible defaults using
\link{ggplot}
}
\author{
Cornel M. Pop
}
