% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surf_shortestPath.R
\name{mesh_to_graph}
\alias{mesh_to_graph}
\title{Transform mesh into an undirected igraph object}
\usage{
mesh_to_graph(mesh)
}
\arguments{
\item{mesh}{a mesh3d object}
}
\value{
A list containing a) "graph", an igraph object; b) "gVertices", a
numeric vector of graph vertices; and c) "edge.curv", a numeric vector
of edge curvature values, computed by adding values for the defining vertices
}
\description{
Transforms a given mesh into an undirected igraph object for
shortest paths and other queries.
}
\note{
This is a very simple function which creates an igraph object from a mesh
and computes edge weights on the basis of vertex curvature values. The only
reason this has been implemented as a separate function is that it is cached,
meaning that repeated calls on the same object (e.g. from sPathQuery) are
fast.
}
\section{TODO}{
 Write unit tests + solve the meanvb versus RMSvb issue.
}

\examples{
data(demoSphere)
graphObj<-Lithics3D:::mesh_to_graph(demoSphere)
}
\author{
Cornel M. Pop
}
\keyword{internal}
