% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transMeshIT.R
\name{mesh_translate_it}
\alias{mesh_translate_it}
\title{Translate vertex IDs to coordinates in the mesh triangle definitions}
\usage{
mesh_translate_it(mesh)
}
\arguments{
\item{mesh}{mesh3d, or mesh3d-like object}
}
\value{
data.frame object
}
\description{
Translates triangle definitions (mesh3d$it) to a dataframe
containing vertex coordinates instead of vertex IDs.
}
\note{
Only the first three rows of the mesh3d$it and mesh3d$vb matrices are
used at the moment. For each row in the mesh3d$it matrix, which contains a
list of column IDs from the mesh3d$vb matrix, three columns are created, in
input order, for a total of 9 columns. These columns contain the row data
from the mesh3d$vb matrix for the specified columns.
Note that value types are not modified (e.g. double in the input == double
in the output)
}
\examples{
data(demoSphere)
coords <- mesh_translate_it(demoSphere)
head(coords)
}
