% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transMeshIT.R
\name{mesh_triangles}
\alias{mesh_triangles}
\title{Produce a list of mesh triangles with vertex coordinates}
\usage{
mesh_triangles(mesh)
}
\arguments{
\item{mesh}{mesh3d, or mesh3d-like object}
}
\value{
list object
}
\description{
Returns a list of triangles defined by 3 vertices (v1, v2, v3).
}
\examples{
data(demoSphere)
triangles <- mesh_triangles(demoSphere)
triangles[[1]]
}
