% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_volume.R
\name{mesh_volume}
\alias{mesh_volume}
\title{Computes mesh volume}
\usage{
mesh_volume(mesh, par_exec = FALSE, max_cores = NA)
}
\arguments{
\item{mesh}{a closed mesh3d object for which volume will be computed}

\item{par_exec}{boolean indicating whether par_exec processing is to be used.
Default value is set to FALSE}

\item{max_cores}{maximum number of cores to use if parallel processing is
requested. Requesting more cores than are physically available will simply
result in all cores being used.}
}
\value{
numeric value of the computed volume or NA in case of bad input
}
\description{
Calculate volume of an arbitrary closed mesh by adding up
signed tetrahedron volumes for each constituent triangle. Please
ensure that the mesh is clean (e.g. no non-manifold vertices, etc)
}
\note{
For small meshes it may be faster to use a single thread due to the
overhead associated with starting the processing cluster.
}
\examples{
data(demoFlake1)
mesh_volume(demoFlake1$mesh)
}
