% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_p2pIntersect.R
\name{p2p.line}
\alias{p2p.line}
\title{Line of intersection between two planes}
\usage{
p2p.line(p1c, p2c)
}
\arguments{
\item{p1c}{Coefficients for the first plane equation}

\item{p2c}{Coefficients for the second plane equation}
}
\value{
A dataframe with the line equation coefficients (vector form)
}
\description{
Compute line of intersection between two non-parallel planes
}
\note{
This works by finding a common point with a third plane whose normal
is the intersection line, and using the p2p.point solver to get its
coordinates. For testing, see applet at:
http://www.songho.ca/math/plane/plane.html#example_2planes
}
\section{TODO}{
 Test function to make sure it operates fine.
}

\examples{
# A lineintersection case
c.m = rbind(c(-7, 1, -2, 3), c(7, -2, 3, -4), c(-5, 3, -6, 10))
res = Lithics3D:::p2p.line(c.m[1,], c.m[2,]) # Use first two planes
\dontrun{
# First point defining the intersection line
p1 = unlist(res[,1:3])
# Second point defining the intersection line
p2 = unlist(res[,1:3] - res[,4:6]*0.1)
library(rgl)
points3d(rbind(c(-5,-5,-5), c(5,5,5)))
lines3d(rbind(p1, p2), lwd=5, col="red") # Draw the line
# Render the intersecting planes
planes3d(c.m[,1], c.m[,2], c.m[,3], c.m[,4], col="green") 
}
}
\author{
Cornel M. Pop
}
