% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pathLength.R
\name{pathLength}
\alias{pathLength}
\title{Measure the length of a path.}
\usage{
pathLength(coords)
}
\arguments{
\item{coords}{An Nx2 or Nx3 matrix-like object containing ordered 2D or
3D coordinates.}
}
\value{
The total length of the path in mesh units
}
\description{
A simple function that adds the distances of individual segments
that form the given input path
}
\section{TODO}{
 Write unit tests. Easy (document bad cases too)
}

\examples{
# A 2D example:
coords = data.frame(x=c(1,1,10), y=c(10,1,1))
pathLength(coords)
}
\author{
Cornel M. Pop
}
