% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pathResample.R
\name{pathResample}
\alias{pathResample}
\title{Resample a curve}
\usage{
pathResample(coords, value, method = "npts")
}
\arguments{
\item{coords}{A NxK matrix-like object containing xy(z) point coordinates,
one per row.}

\item{value}{A numeric value to use for sampling given the requested method.}

\item{method}{The sampling method to use. Valid options are: a) "npts" which
requests the number of points specified in the value parameter and is the
default; and b) "dist", which specifies the requested distance between the
points. Note that distance is computed by summing segment lengths, so the
distance between the sampled points in space will NOT be the same.}
}
\value{
A resampled curve as output by geomorph::digit.curves
}
\description{
Resamples a curve, defined by an ordered set of coordinates, by
the desired number of points or desired distance between points.
}
\note{
This is a wrapper for geomorph::digit.curves, and parameters are passed
directly to that function without any checking. Ergo that function is
responsible for error handling.
}
\section{TODO}{
 Write tests. ALSO, check length of output!!
}

\examples{
print("Coming soon")
}
\author{
Cornel M. Pop
}
