% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_p2pIntersect.R
\name{planeCoefs}
\alias{planeCoefs}
\title{Get plane equation coefficients from 3 points:}
\usage{
planeCoefs(p)
}
\arguments{
\item{p}{A 3x3 matrix-like object of point coordinates, one point per row.}
}
\value{
A vector of scalar plane equation coefficients in the form
Ax + By + Cz + D = 0
}
\description{
Computes the coefficients of the plane equation for a plane
defined by 3 3D points.
}
\section{TODO}{
 Write unit tests
}

\examples{
p = data.frame(x=c(1,4,5), y=c(0,5,1), z=c(1.4, -4, 4))
res = planeCoefs(p)
\dontrun{
library(rgl)
points3d(p, size=10, col="red")
planes3d(res[1], res[2], res[3], res[4], col="green")
}
}
\author{
Cornel M. Pop
}
