% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_mark_pois.R
\name{proj_poi}
\alias{proj_poi}
\title{Determine Point of Interest (POI) on a mesh surface using ray tracing}
\usage{
proj_poi(ray, mesh)
}
\arguments{
\item{ray}{A 2x3 matrix-like object containing x, y, and z coordinates
for two points that define the ray.}

\item{mesh}{A triangular mesh object (\code{mesh3d}).}
}
\value{
A vector with the x, y, and z coordinates of the POI.
}
\description{
Determines the location of a POI on the mesh surface based on a ray. This
function is used internally by the \code{\link[=mesh_mark_pois]{mesh_mark_pois()}} function,
and may not be very useful in other contexts.
}
\details{
The function uses ray tracing to project the given ray onto the
mesh surface and finds the closest intersection point. If the ray does not
intersect with the mesh, an empty object is returned.
}
\seealso{
\code{\link[=rayTrace]{rayTrace()}} for the ray tracing implementation.
\code{\link[=mesh_mark_pois]{mesh_mark_pois()}} for ray determination.
}
\keyword{internal}
