% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_axis.R
\name{rotatePt_2D}
\alias{rotatePt_2D}
\title{Rotate a 2D point}
\usage{
rotatePt_2D(x, y, angle)
}
\arguments{
\item{x}{X-axis coordinate of the point to be rotated}

\item{y}{Y-axis coordinate of the point to be rotated}

\item{angle}{Rotation angle, given in radians}
}
\value{
A vector containing the rotated coordinates (x,y)
}
\description{
Rotates a 2D point by a given angle (counter-clockwise)
}
\examples{
x = 1.5
y = 0
a = 0.2
pt.r = rotatePt_2D(x, y, a) # 1.4701, 0.298004
}
\author{
Cornel M. Pop
}
