% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_cutPlane.R
\name{split_pts}
\alias{split_pts}
\title{Split pointcloud by plane}
\usage{
split_pts(coords, p)
}
\arguments{
\item{coords}{A matrix-like object with at least 3 columns representing
N x 3 XYZ coordinates. Additional columns will be ignored.}

\item{p}{A 3x3 matrix-like object with coordinates defining a plane, one per
row.}
}
\value{
A list of two logical vectors of length N, reporting if the
coordinate falls above the plane, or not (upr), and if the coordinate falls
below the plane, or not (lwr). These two vectors are not exact mirrors of
each other since some coordinates may fall on the plane.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Splits the input coordinates by a 3D plane. This function
produces identical results to \link[Morpho:cutSpace]{cutSpace}, but should run
substantially faster.
}
\examples{
library(rgl)
data(demoFlake2)
mesh.o <- Morpho::pcAlign(demoFlake2$mesh)
mpts <- t(mesh.o$vb)
p <- data.frame(x = c(0, 1, 1),
                y = c(0, 1, 0.5),
                z = c(0, 1, 0))
res <- split_pts(mpts, p)
points3d(mpts[res$upr, 1:3], color="red")
points3d(mpts[!res$upr, 1:3], color="blue")
}
\seealso{
\link[Morpho:cutSpace]{cutSpace}
}
\author{
Cornel M. Pop
}
