% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_rotate.R
\name{alignAxis}
\alias{alignAxis}
\title{Align to axis (obsolete)}
\usage{
alignAxis(l, coords)
}
\arguments{
\item{l}{A 2X3 matrix-like object containing xyz point coordiantes defining
the alignment vector, one coordinate per row.}

\item{coords}{A Nx3 matrix-like object containing xyz point coordinates, one
per row.}
}
\value{
A list containing rotated coordinates (coords) and the rotated
input vector (l).
}
\description{
This function is obsolete. Please use the more general
\link{rotate_v2v} function instead.
}
\author{
Cornel M. Pop
}
