% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeAngles_yw.R
\name{.cluster_patch}
\alias{.cluster_patch}
\title{Assign patch coordinates to one of two surfaces}
\usage{
.cluster_patch(coords, normals, rad, lambda)
}
\arguments{
\item{coords}{An Nx3 numeric matrix-like object where each row represents the
x, y, and z coordinates of a mesh vertex.}

\item{normals}{An Nx3 numeric matrix-like object where each row represents
the x, y, and z components of a normal vector corresponding to each vertex in
\code{coords}.}

\item{rad}{A positive numeric value representing the radius that influences
the classification.}

\item{lambda}{A positive numeric value used as a scaling factor in the
classification calculation.}
}
\value{
A numeric vector where each element corresponds to a point in
\code{coords} and indicates the classification of that point into one of two
groups (1 or 2).
}
\description{
Classifies 3D coordinates of mesh vertices into one of two groups
representing intersecting surfaces.
}
\details{
This is a translation of c++ code from the Meshlab Virtual Goniometer plugin
developed by the \href{http://amaaze.umn.edu}{AMAAZE consortium}
}
\examples{
coords <- matrix(rnorm(300), ncol = 3)
normals <- matrix(rnorm(300), ncol = 3)
rad <- 1
lambda <- 0.5
classification <- .cluster_patch(coords, normals, rad, lambda)
print(classification)
}
\keyword{internal}
