% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeAngles_yw.R
\name{.PCA_smallest_eig}
\alias{.PCA_smallest_eig}
\title{Compute smallest eigenvalue and its corresponding eigenvector}
\usage{
.PCA_smallest_eig(xyz_data, C, c, center = FALSE)
}
\arguments{
\item{xyz_data}{A Nx3 numeric matrix-like object where each row represents
the x, y, and z coordinates of a point in space.}

\item{C}{A numeric vector indicating the group of each observation in
\code{xyz_data}.}

\item{c}{A numeric value specifying the group to be processed.}

\item{center}{Logical; if \code{TRUE}, the data will be centered (mean subtracted)
before computing the covariance matrix. Defaults to \code{FALSE}.}
}
\value{
A list containing two elements: \code{smallest_eigenvalue}, the smallest
eigenvalue of the covariance matrix, and \code{smallest_eigenvector}, the
corresponding eigenvector.
}
\description{
Calculates the smallest eigenvalue and its corresponding eigenvector of the
covariance matrix of a filtered 3D coordinates data set, optionally centering
the latter.
}
\details{
This is a translation of c++ code from the Meshlab Virtual Goniometer plugin
developed by the \href{http://amaaze.umn.edu}{AMAAZE consortium}
}
\examples{
xyz_data <- matrix(rnorm(300), ncol = 3)
C <- sample(1:2, 100, replace = TRUE)
result <- .PCA_smallest_eig(xyz_data, C, 1)
print(result)
}
\keyword{internal}
