% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surf_edgeIntersects.R
\name{e2sIntersect}
\alias{e2sIntersect}
\title{Mesh edge / sphere intersection}
\usage{
e2sIntersect(e.ids, s, mpts, medges)
}
\arguments{
\item{e.ids}{A vector containing edge IDs as output by Rvcg
vcgGetEdge function with "unique" set to TRUE.}

\item{s}{A vector of length 4 (center.x, center.y, center.z, radius) defining
a sphere}

\item{mpts}{Nx4 matrix-like object corresponding to the transposed mesh$vb}

\item{medges}{An Nx4 dataframe corresponding to the output of
Rvcg::vcgGetEdge(mesh, unique=T).}
}
\value{
A data.frame with intersection coordinates (x,y,z,x1,y1,z1), one per
row, for all input edges, where the row names correspond to input edge ids.
NA will be returned in cases where there are no intersections. The first
coordinate (x,y,z) will always be located ON the segment and farthest
from the center of the sphere (i.e. if the segment crosses the entire sphere).
}
\description{
Computes the intersection coordinates of the input mesh edges
and a sphere.
}
\section{TODO}{
 Test this function with 3 different scenarios: no inters,
inters, and both inters on segment.
}

\examples{
print("TODO")
}
