% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeAngles_visualize.R
\name{edge_angles_vis2d}
\alias{edge_angles_vis2d}
\title{2D visualizer for edgeAngles output}
\usage{
edge_angles_vis2d(ea_output, mesh, ncol = 2, nrow = 2)
}
\arguments{
\item{ea_output}{A list corresponding to the unmodified output of
\code{\link[=edgeAngles]{edgeAngles()}}.}

\item{mesh}{An \code{mesh3d} object corresponding to the mesh on which the angles
were measured.}

\item{ncol}{A number indicating the number of sub-plots to fit horizontally
in each output figure.}

\item{nrow}{A number indicating the number of sub-plots to fit vertically
in each output figure.}
}
\value{
A list of figures containing subplots for each angle measurement;
each is labelled according to the index of the measurement along the edge.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Shows how edge angles were measured in a previous run of the \code{\link[=edgeAngles]{edgeAngles()}}
function.
}
\details{
Loops through the output of \code{\link[=edgeAngles]{edgeAngles()}} and plots key parameters for each
angle measurement. The number of sub-plots per figure will be computed based
on the \code{ncol} and \code{nrow} parameters. If you would like individual figures
for each sub-plot, simply set both \code{ncol} and \code{nrow} to 1.

Note that all sub-plots from given output figure will be drawn to the same
scale, but different figures will have different scales.
}
\examples{
\dontrun{
# Compute edge angles using included demo data:
data(demoFlake2)
e.curve = sPathConnect(demoFlake2$lms[1:4, ],
                       demoFlake2$mesh, path.choice = "ridges")
mv <- t(demoFlake2$mesh$vb)
path.res <- pathResample(mv[e.curve,1:3], 30, method = "npts")
res = edgeAngles(demoFlake2$mesh, path.res, m.d = 3)

# Generate figures:
angle_plots <- edge_angles_vis2d(res,
                                 mesh = demoFlake2$mesh,
                                 ncol = 3)
# See the first figure:
angle_plots[[1]]
}
}
\author{
Cornel M. Pop
}
