% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_splitbypath.R
\name{mesh_segment_by_path}
\alias{mesh_segment_by_path}
\title{Section mesh using a path}
\usage{
mesh_segment_by_path(mesh, lms, path.choice = "ridges")
}
\arguments{
\item{mesh}{A (clean) mesh3d object}

\item{lms}{An ordered set of coordinates}

\item{path.choice}{TODO: Described in sPathConnect.}
}
\description{
Cuts a mesh along a closed path defined on the basis of an
ordered set of coordinates (ideally landmarks), and splits the mesh into
multiple components. Note that you must provide a clean mesh, so please use
\link{vcgClean} on the input.
}
\note{
Minimum output mesh segment size is 3 faces. Anything smaller gets
discarded silently
}
\examples{
library(rgl)
zzz <- mesh_segment_by_path(demoFlake2$mesh, demoFlake2$lms)
shade3d(zzz[[1]], color="green")
shade3d(zzz[[2]], color="blue")
}
\author{
Cornel M. Pop
}
