% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_dist.R
\name{ptOnSeg}
\alias{ptOnSeg}
\title{Determine if coordinates are on a line segment (Deprecated)}
\usage{
ptOnSeg(coords, l, d.t = 0)
}
\arguments{
\item{coords}{A Nx3 matrix-like object with point coordinates, one per row.}

\item{l}{A 2x3 matrix-like object with coordinates defining a line segment,
one coordinate per row.}

\item{d.t}{(optional) Tolerance value for distance from the line segment;
useful to deal with floating point precision errors. Default is 0.}
}
\value{
A boolean vector indicating, for each input coordinate, whether they
lie on the given line segment
}
\description{
This function has been renamed to \link{coords_onseg}.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\author{
Cornel M. Pop
}
\keyword{internal}
