% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surf_shortestPath.R
\name{sPathConnect}
\alias{sPathConnect}
\title{Connect landmarks by the shortest geodesic path}
\usage{
sPathConnect(coords, mesh, path.choice = "any")
}
\arguments{
\item{coords}{data.frame-like object of ordered 3D coordinates (one per row)}

\item{mesh}{a mesh3d object}

\item{path.choice}{path choice specification. Options are the same as for
the sPathQuery function.}
}
\value{
A numeric, ordered list of all vertex IDs in the path (includes
start and end vertices)
}
\description{
Projects the given set of ordered coordinates onto the mesh and
connects them following the shortest path (weighted or unweighted) along
triangle edges (i.e. connected vertices) on a target mesh. The curve
described by the coordinates may be open or closed.
}
\note{
Internally, this function uses sPathQuerry and mapOnMesh, so
note that the input coordinates are not mapped onto the closest surface
point, but to the nearest mesh vertex. This may lead to small
discrepancies with input landmarks that are already mapped onto the mesh
surface. See mapOnMesh documentation for more information.
}
\examples{
# TODO!
}
\seealso{
sPathQuery
}
